Moved MapsDataClient to Common
