+++ shared/src/commonMain/kotlin/com/stephenwoerner/dubvtransittwo/shared/directions/KDirectionsApi.kt
@@ -1,0 +1,4 @@
+package com.stephenwoerner.dubvtransittwo.shared.directions
+
+class KDirectionsApi {
+}
