+++ shared/src/iosMain/kotlin/com/stephenwoerner/dubvtransittwo/shared/directions/KDirectionsApi.kt
@@ -1,0 +1,7 @@
+package com.stephenwoerner.dubvtransittwo.shared.directions
+
+actual class KDirectionsApi {
+    actual fun newRequest(): StepsAndDuration {
+        return StepsAndDuration(arrayListOf<SimpleDirections>(), 0,false)
+    }
+}
